- Made commands delete the original message (only when successful).
# Commands

- nux autobump
Automatically uses /bump in the given channel ID, with varying time from 2 hours to 2 and a half hours.

- nux autoowod
Automatically uses "owo daily" in the given channel ID, with the time that you give it and varying delay so it looks natural.

- nux voicewatch
Logs voice channels, works just like `nux watch`.

- nux addall (--force)
Adds everyone in the given server ID, but doesn't work without --force.

- nux autoreact
Automatically reacts with an emoji you give in a channel provided ID with the keyword that triggers it.